Refactor DepositAddresForm, implement properly folderization to have readable code
