+++ components/Swap/Form/DepositAddressForm/EasyDepositBanner.tsx
