+++ components/Swap/Form/ExchangeForm.tsx
@@ -31,11 +31,9 @@ import shortenString from "@/components/utils/ShortenString";
 
 type Props = {
     partner?: Partner;
-    showBanner: boolean;
-    dismissBanner: () => void;
 };
 
-const ExchangeForm: FC<Props> = ({ partner, showBanner, dismissBanner }) => {
+const ExchangeForm: FC<Props> = ({ partner }) => {
     const {
         values, isSubmitting
     } = useFormikContext<SwapFormValues>();
