+++ components/Swap/Withdraw/Processing/Processing.tsx
@@ -349,7 +349,7 @@ const Processing: FC<Props> = ({ swapBasicData, swapDetails, quote, refuel }) =>
 
     return (
         <Widget.Content fitContent>
-            <div className={`w-full min-h-102.5 h-full space-y-3 flex flex-col justify-between text-primary-text`}>
+            <div className={`w-full min-h-102.5 h-full space-y-2 flex flex-col justify-between text-primary-text`}>
                 <SwapSummary />
                 <div className="bg-secondary-500 font-normal px-3 pt-6 pb-3 rounded-2xl space-y-4 flex flex-col w-full relative z-10 divide-y-2 divide-secondary-300 divide-dashed">
                     <div className='pb-4'>
